add number of seats input
